pyproject.toml
[tool.poetry]
name = "tiny-mall"
version = "0.1.0"
description = ""
authors = ["Kiyan Gauss <fourcels@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.14.0"
SQLAlchemy = "^1.4.25"
psycopg2 = "^2.9.1"
python-jose = "^3.3.0"
python-dotenv = "^0.19.1"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
